+++ sdkjs-plugins/v1/onlyoffice-types/src/plugin/asc-plugin.d.ts
@@ -1,0 +1,1 @@
+export type { AscPlugin } from "../../index";