+++ sdkjs-plugins/v1/onlyoffice-types/src/plugin/asc.d.ts
@@ -1,0 +1,6 @@
+export type {
+    Asc,
+    AscPlugin,
+    AscTheme,
+    PluginScope,
+} from "../../index";