+++ sdkjs-plugins/v1/onlyoffice-types/src/plugin/index.d.ts
@@ -1,0 +1,26 @@
+export {
+    type Asc,
+    type AscPlugin,
+    type AscTheme,
+    type PluginScope,
+} from "./asc";
+export {
+    type PluginEventMap,
+    type PluginEventHandler,
+    type ContextMenuShowEvent,
+    type PluginEventName,
+    type PluginEventCallback,
+    type PluginEditorEventName,
+    type PluginEditorEventCallback,
+} from "./events";
+export {
+    type ButtonBase,
+    type ButtonContextMenu,
+    type ButtonToolbar,
+    type ButtonContentControl,
+    type ButtonWindowHeader,
+    type ButtonMenuItem,
+    type Buttons,
+    type WindowHeaderFrameOptions,
+} from "./menus";
+export { type PluginWindow } from "./windows";