+++ sdkjs-plugins/v1/onlyoffice-types/src/plugin/menus.d.ts
@@ -1,0 +1,10 @@
+export type {
+    ButtonBase,
+    ButtonContextMenu,
+    ButtonToolbar,
+    ButtonContentControl,
+    ButtonWindowHeader,
+    ButtonMenuItem,
+    Buttons,
+    WindowHeaderFrameOptions,
+} from "../../index";