+++ sdkjs-plugins/v1/onlyoffice-types/src/services/index.d.ts
@@ -1,0 +1,2 @@
+export { type AscDesktopEditor } from "./desktop-editor";
+export { type AscSimpleRequest, type AscSimpleRequestOptions } from "./simple-request";